Fix bill cron job

+++ src/pages/api/tenancyModification/property.ts
@@ -23,29 +23,82 @@ export default async function handler(
 
   const decodedToken = authenticateUser(req, res);
 
+  if (!decodedToken?.user?._id) {
+    return res.status(401).json({
+      success: false,
+      msg: "Not Authorized",
+    });
+  }
+
   try {
     await mongoose.connect(process.env.NEXT_PUBLIC_MONGODB_SRV || "mongodb://localhost:27017/test_db").then(() => {
       // USER
       const { _id, role } = decodedToken.user;
 
+      //type of request
       const { method } = req;
+
+      // Property ID
+      const property = req.body.property || req.query.id
+
+      // Reject if no property provided
+      if (!property) {
+        return res.status(400).json({
+          success: false,
+          msg: "No property provided",
+        });
+      }
+
+      // Get permissions
+      const userRoles = decodedToken.userRoles
+      const userPropertyRoles = userRoles?.find((role: any) => role.property === property)
+      const permissions = userPropertyRoles?.role?.permissions
+
       switch (method) {
         case "GET":
-        //   if (id) {
-        //     return fetchSingleUnit(req, res)
-        //   }
+          const getPermission = permissions?.find((permission: any) => permission.name === "View tenancy modification")
+          if (!getPermission) {
+            return res.status(401).json({
+              success: false,
+              msg: "Not Authorized",
+            });
+          }
+          //   if (id) {
+          //     return fetchSingleUnit(req, res)
+          //   }
           fetchTenancyModificationsByProperty(req, res);
           break;
-          // case "GET":
-          //   fetchSingleUnit(req, res);
+        // case "GET":
+        //   fetchSingleUnit(req, res);
         //   break;
         case "POST":
+          const postPermission = permissions?.find((permission: any) => permission.name === "Create tenancy modification")
+          if (!postPermission) {
+            return res.status(401).json({
+              success: false,
+              msg: "Not Authorized",
+            });
+          }
           createTenancyModification(req, res);
           break;
         case "PUT":
+          const putPermission = permissions?.find((permission: any) => permission.name === "Edit tenancy modification")
+          if (!putPermission) {
+            return res.status(401).json({
+              success: false,
+              msg: "Not Authorized",
+            });
+          }
           updateUnit(req, res);
           break;
         case "DELETE":
+          const deletePermission = permissions?.find((permission: any) => permission.name === "Delete tenancy modification")
+          if (!deletePermission) {
+            return res.status(401).json({
+              success: false,
+              msg: "Not Authorized",
+            });
+          }
           deleteUnit(req, res);
           break;
         default:
